scraping/requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
Pillow==10.4.0

